requirements.txt
fastapi
uvicorn[standard]
streamlit
redis
fastf1
pandas
numpy
scikit-learn
torch
pydantic
python-dotenv
tqdm
requests